Split up files and updated namespaces.
